Update slice-update with return code and UID check.
This change modifies slice-update to allow running smoothly from
anisble. In particular, the  background process used to trigger the
slice-update is disowned to allow a clean exit when called remotely.

This change also includes a UID check for root, since non-root users
cannot view the content of the /vsys directory.